requirements.txt
chromadb==0.4.22
markdown==3.5.2
python-frontmatter==1.1.0
tqdm==4.66.1
numpy<2.0.0
scikit-learn>=1.3.0
beautifulsoup4==4.12.3
openai>=1.12.0
python-dotenv==1.0.1
sentence-transformers>=2.5.1
